moved the main of server to click package
